requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.22.0
pydantic>=2.8.0,<3.0.0
python-multipart==0.0.9
python-socketio
grpcio

passlib==1.7.4
passlib[bcrypt]
PyJWT[crypto]

requests==2.32.2
aiohttp==3.9.5
httpx

# AI libraries
openai==1.68.2
anthropic==0.48.0
google-generativeai==0.8.3
vertexai

# Database
pymongo
peewee
SQLAlchemy
boto3
redis
sqlmodel==0.0.23
chromadb==0.6.2
psycopg2-binary

# Observability
langfuse==2.60.1
ddtrace

# ML libraries
torch
numpy
pandas

xgboost
scikit-learn

# NLP libraries
sentence-transformers
transformers
tokenizers
nltk
tiktoken

# Image processing
Pillow
opencv-python

# Visualization
matplotlib
seaborn

# Web scraping
selenium
playwright
beautifulsoup4

# Llama Index for RAG
llama-index==0.12.25
llama-index-llms-ollama

#spark
pyspark==3.5.5

#langchain
langchain==0.3.21
langchain-community==0.3.20
langchain-core==0.3.49
langchain-ollama==0.3.0
langchain-text-splitters==0.3.7
langchain_openai=0.3.11

